Add files via upload
PROBLEM STATEMENT

The purpose is to recognize the emotion and affective state of the speaker from his/her speech signal.

DATA SOURCE USED

We have used the RAVDESS dataset in this project. It is one of the more common datasets used for this exercise by others. It's well-liked because of its quality of speakers, and recording and it has 24 actors of different genders. Here are the filename identifiers as per the official RAVDESS website: